requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
SQLAlchemy>=2.0,<3.0
PyMySQL>=1.1,<2.0
Jinja2>=3.1,<4.0
pydantic>=2.11,<3.0
python-dotenv>=1.0,<2.0
